Book Database - GraphQL
A GraphQL project used to create and store book titles and author names. The database is implemented through Mongoose and Mongo DB. The temp files contains React files for the project, implementing the Apollo Client to reach into GraphQL and the server side. Construction of this side is still in progress.