Book Database - GraphQL Project
A server-side GraphQL project that logs books and authors and retrieves them from a MongoDB database using Mongoose. Implementation for the client side is still in progress using the React library and Apollo Client, but files for the client side can be found within the temp folder. The GraphQL serverside database is accessed through the GraphiQL interface.